refactor: util, strategy 패키지 삭제
