+++ src/test/java/racingcar/domain/CarsTest.java
@@ -9,8 +9,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import racingcar.domain.util.CarFactory;
-
 public class CarsTest {
     private Car forky;
     private Car kun;
