+++ src/test/java/racingcar/domain/MovePredicateTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain.strategy;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
